The "persistent" polling option now prevents channels from being destroyed
This is mainly to circumvent a bug where Hookbox continues polling after
the channel is destroyed, which can result in an enormous amount of
requests in the long run.

+++ hookbox/channel.py
@@ -324,7 +324,7 @@ class Channel(object):
             self.history.append(('UNSUBSCRIBE', frame))
             self.prune_history()
         
-        if not self.subscribers:
+        if not self.subscribers and self._options['poll']['mode'] != 'persistent':
             self.server.destroy_channel(self.name)
     
     def destroy(self, needs_auth=True):
